move all i2c code into the serial package
