+++ components/movementsensor/gpsrtkserial/correction_writer.go
@@ -1,7 +1,7 @@
 //go:build linux
 
 // Package gpsrtkpmtk implements a gps using serial connection
-package gpsrtkpmtk
+package gpsrtkserial
 
 import (
 	"context"
