+++ components/movementsensor/gpsrtkserial/gpsrtkpmtk_nonlinux.go
@@ -1,2 +1,2 @@
 // Package gpsrtkpmtk is unimplemented for non-Linux OSes.
-package gpsrtkpmtk
+package gpsrtkserial
