+++ components/movementsensor/register/register.go
@@ -7,7 +7,6 @@ import (
 	_ "go.viam.com/rdk/components/movementsensor/dualgps"
 	_ "go.viam.com/rdk/components/movementsensor/fake"
 	_ "go.viam.com/rdk/components/movementsensor/gpsnmea"
-	_ "go.viam.com/rdk/components/movementsensor/gpsrtkpmtk"
 	_ "go.viam.com/rdk/components/movementsensor/gpsrtkserial"
 	_ "go.viam.com/rdk/components/movementsensor/imuvectornav"
 	_ "go.viam.com/rdk/components/movementsensor/imuwit"
